authentication_service/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
jwt==1.3.1
pycparser==2.21
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
